Move pressure and moveable out of tile.py
