+++ src/collision_handlers/polling_collision_handler.py
@@ -10,7 +10,7 @@ import pygame
 import src.game_object.bomb as bomb
 import src.game_object.tile as tile
 import src.graphics as graphics
-from src.game_object.tile import PressurePlate
+from src.game_object.pressure_plate import PressurePlate
 from src.game_object.tile import Triggerable
