+++ src/level_editor.py
@@ -1,6 +1,7 @@
 import src.game_object.bomb as bomb
 import src.game_object.tile as tile
 from src.game_object.switch_tile import Switch
+from src.game_object.pressure_plate import PressurePlate
 from src.game_object.portal import Portal
 import src.game_object.actor as actor
 from pprint import pprint
@@ -161,7 +162,7 @@ class LevelData():
         """Makes sure that the switches are passed to the correct
            door object(Triggerable)"""
         for state in self.sprites:
-            if isinstance(state, Switch) or isinstance(state, tile.PressurePlate):
+            if isinstance(state, Switch) or isinstance(state, PressurePlate):
                 for trigger in self.sprites:
                     if isinstance(trigger, tile.Triggerable):
                         if state.triggers == trigger.triggered_id:
