+++ src/tile_factory.py
@@ -1,9 +1,9 @@
 from src.game_object.tile import Triggerable
-from src.game_object.tile import MoveableTile
+from src.game_object.moveable_tile import MoveableTile
 from src.game_object.finish_tile import FinishTile
 from src.game_object.switch_tile import Switch
 from src.game_object.portal import Portal
-from src.game_object.tile import PressurePlate
+from src.game_object.pressure_plate import PressurePlate
 from src.game_object.tile import Tile
 from src.game_object.bomb import Bomb
 from src.game_object.actor import Actor
